+++ package.json
@@ -53,6 +53,8 @@
 		"geth-binary:start": "WEB3_SYSTEM_TEST_BACKEND=geth && ./scripts/geth_binary.sh start",
 		"geth-binary:start:background": "WEB3_SYSTEM_TEST_BACKEND=geth && ./scripts/geth_binary.sh start 1",
 		"geth-binary:stop": "WEB3_SYSTEM_TEST_BACKEND=geth && ./scripts/geth_binary.sh stop",
+		"geth-manual:start:background": "WEB3_SYSTEM_TEST_BACKEND=geth && ./scripts/geth_binary.sh manualStart 1",
+		"geth-manual:stop": "WEB3_SYSTEM_TEST_BACKEND=geth && ./scripts/geth_binary.sh stop",
 		"lint": "lerna run lint --stream --parallel",
 		"lint:fix": "lerna run lint:fix --stream --parallel",
 		"format": "lerna run format --stream --parallel && prettier --write ./scripts/**/*.ts",
@@ -61,6 +63,7 @@
 		"test:coverage:integration": "lerna run test:coverage:integration --stream --parallel",
 		"test:unit": "lerna run test:unit --stream --parallel && jest --config=./scripts/jest.config.js",
 		"test:integration": "lerna run test:integration --stream",
+		"test:manual": "lerna run test:manual --stream",
 		"test:e2e:ganache:http": "./scripts/test-runner.sh ganache http",
 		"test:e2e:ganache:ws": "./scripts/test-runner.sh ganache ws",
 		"test:e2e:geth:http": "./scripts/test-runner.sh geth http",
@@ -74,6 +77,8 @@
 		"test:e2e:mainnet:ws": "./scripts/test-runner.sh mainnet ws",
 		"test:e2e:sepolia:http": "./scripts/test-runner.sh sepolia http",
 		"test:e2e:sepolia:ws": "./scripts/test-runner.sh sepolia ws",
+		"test:e2e:manual:geth:ws": "./scripts/test-runner.sh geth-manual ws node manual",
+		"test:e2e:manual:geth:ipc": "./scripts/test-runner.sh geth-manual ipc node manual",
 		"generate:accounts": "node ./scripts/gen_accounts.js",
 		"pre-blackbox": "yarn config set registry http://localhost:4873 && git init && git config --global user.email \"ci@github.com\" && git config --global user.name \"CI\"",
 		"post-blackbox": "./scripts/verdaccio.sh stop",
@@ -85,10 +90,6 @@
 		"test:blackbox:geth:ws": "yarn pre-blackbox && yarn geth:start:background && ./scripts/verdaccio.sh startBackgroundAndPublish && lerna run test:blackbox:geth:ws --stream && yarn post-blackbox:geth",
 		"test:blackbox:infura:http": "yarn pre-blackbox && ./scripts/verdaccio.sh startBackgroundAndPublish && lerna run test:blackbox:infura:http --stream && yarn post-blackbox",
 		"test:blackbox:infura:ws": "yarn pre-blackbox && ./scripts/verdaccio.sh startBackgroundAndPublish && lerna run test:blackbox:infura:ws --stream && yarn post-blackbox",
-		"test:manual:stress:data": "packages/web3/test/stress/large_request_data/start.sh",
-		"test:manual:stress:requests": "packages/web3/test/stress/many_requests/start.sh",
-		"test:manual:stress:validation": "npx ts-node packages/web3/test/stress/large_validator_data/index.ts",
-		"test:manual:stress": "yarn test:manual:stress:data && yarn test:manual:stress:validation && yarn test:stress:requests",
 		"husky:install": "husky install",
 		"husky:uninstall": "husky uninstall",
 		"postinstall": "yarn build",
