+++ packages/web3/package.json
@@ -53,6 +53,7 @@
 		"test:watch": "npm test -- --watch",
 		"test:unit": "jest --config=./test/unit/jest.config.js",
 		"test:integration": "jest --config=./test/integration/jest.config.js --forceExit",
+		"test:manual": "jest --config=./test/manual/jest.config.js --forceExit",
 		"test:blackbox:ganache:http": "./scripts/black_box_test.sh ganache http",
 		"test:blackbox:ganache:ws": "./scripts/black_box_test.sh ganache ws",
 		"test:blackbox:geth:http": "./scripts/black_box_test.sh geth http",
