+++ scripts/geth_binary.sh
@@ -70,12 +70,13 @@ start() {
 	fi
 }
 
-startStress() {
+startManual() {
     download
 
-    echo "Starting geth..."
+    echo "Starting manual geth..."
     echo "geth --ipcpath $IPC_PATH --nodiscover --nousb --ws --ws.addr 0.0.0.0 --ws.port $WEB3_SYSTEM_TEST_PORT --http --http.addr 0.0.0.0 --http.port $WEB3_SYSTEM_TEST_PORT --allow-insecure-unlock --http.api personal,web3,eth,admin,debug,txpool,net --ws.api personal,web3,eth,admin,debug,miner,txpool,net --dev --mine --dev.period=0 --dev.gaslimit 9000000000000000 --rpc.txfeecap=1000000 &>/dev/null &"
     ${TMP_FOLDER}/geth --ipcpath $IPC_PATH --nodiscover --nousb --ws --ws.addr 0.0.0.0 --ws.port $WEB3_SYSTEM_TEST_PORT --http --http.addr 0.0.0.0 --http.port $WEB3_SYSTEM_TEST_PORT --allow-insecure-unlock --http.api personal,web3,eth,admin,debug,txpool,net --ws.api personal,web3,eth,admin,debug,miner,txpool,net --dev --mine --dev.period=0 --rpc.enabledeprecatedpersonal --dev.gaslimit 9000000000000000 --rpc.txfeecap=1000000 &>/dev/null &
+
     echo "Waiting for geth..."
     npx wait-port -t 10000 "$WEB3_SYSTEM_TEST_PORT"
 
@@ -107,7 +108,7 @@ stop() {
 }
 
 case $1 in
-stressStart) startStress ;;
+manualStart) startManual ;;
 syncStart) startSync ;;
 syncStop) syncStop ;;
 start) start ;;
